server/requirements.txt
flask
flask-sqlalchemy
flask-limiter
flask-cors
mysql-connector-python
bcrypt
pycryptodome
beautifulsoup4
requests
